backend/requirements.txt
# Web framework
fastapi==0.115.5
starlette==0.41.2
uvicorn[standard]==0.30.1
pydantic==2.7.4
httpx==0.27.2

# Data
polars==1.13.1
pyarrow==17.0.0

# LLM
google-genai>=1.66.0
openai>=1.40.0

# UI
streamlit>=1.38.0

# Config
python-dotenv==1.0.0

# Optional heavy deps (commented out for initial run on Windows without build tools)
# numpy==2.1.3
# scipy==1.14.1
# scikit-learn==1.5.2
# diskcache==5.6.3
# pyarrow==17.0.0
